Update README.md -- IPv6 handling
Docker seems to be utilizing IPv6 by default in a dual stack environment.  This is allowing potential DNS and traffic leaks bypassing the vpn. 

I believe the changes made for the configuration and disabling IPv6 by default on the networks is a good option to solve this.  I am basing this information off of TrueNAS 25.04 docker configuration currently but will be testing more on and Ubuntu 24.04 install as well.